Add-On browser should use 50/50 layout Fixes #636
